+++ src/Function/debounce/debounce.ts
@@ -1,11 +1,16 @@
+type IFunction = {
+	(this: void, ...args: unknown[]): void;
+};
+
 /**
  * 防抖函数
  * @returns
  * @example
  *
  */
-export default function debounce(func: typeof Function, wait: number, immediate = false): typeof Function {
-	const timer = null;
+export default function debounce(func: IFunction, wait: number, immediate = false): IFunction {
+	// eslint-disable-next-line @typescript-eslint/no-explicit-any
+	let timer: any = null;
 	return function (...args: unknown[]): void {
 		if (immediate) {
 			func.apply(this, args);