+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 1.0.4 (2022-01-10)
 
 * 新增Store分类
-* 新增store、cookie、debounce、throttle、getStyle、setStyle方法。
+* 新增store、cookie、debounce、throttle方法。
 * 移除getStore、setStore、removeStore方法。
 
 ## 1.0.0 (2021-12-16)